VCard: Don't copy list on `getProperties(Class)`.
This is what I actually intended to do before deciding I should tackle
https://github.com/fullcontact/ez-vcard/pull/5 first.

Basically, instead of creating a brand new `List` to hold the results
from `getProperties(Class)`, we instead simply return the `List` the
`ListMultimap` gives us, cast to the appropriate type.

This is both better for performance, and means one can do destructive
filtering of properties of a particular class by using
`Iterator.remove()`. Previously, we would accumulate the properties we
wanted to remove into a new `List`, then use `removeProperty` to remove
each of them, which is O(n²) (significant when _n_ = 3500) and a lot
harder to read overall.